+++ src/EliteChroma.Core/Layers/BackgroundLayer.cs
@@ -4,30 +4,53 @@ using Colore.Data;
 using Colore.Effects.Keyboard;
 using EliteChroma.Chroma;
 using EliteChroma.Elite;
+using EliteFiles.Status;
 
 namespace EliteChroma.Core.Layers
 {
     [SuppressMessage("Performance", "CA1812:Avoid uninstantiated internal classes", Justification = "Instantiated by ChromaController.InitChromaEffect().")]
     internal sealed class BackgroundLayer : LayerBase
     {
-        private static readonly TimeSpan _fadeDuration = TimeSpan.FromSeconds(1);
+        private static readonly TimeSpan _fadeDuration = TimeSpan.FromSeconds(0.2);
 
-        private GameProcessState _lastState;
-        private Color _animKbdC1;
-        private Color _animKbdC2;
-        private Color _animDevC1;
-        private Color _animDevC2;
+        private (bool HardpointsDeployed, VehicleMode Mode) _lastState;
+        private AmbientColors _animC1;
+        private AmbientColors _animC2;
+
+        private enum VehicleMode
+        {
+            Combat,
+            Analysis,
+            Landing,
+        }
 
         public override int Order => 0;
 
         protected override void OnRender(ChromaCanvas canvas)
         {
-            if (Game.ProcessState != _lastState)
+            var hardpointsDeployed = Game.Status.HasFlag(Flags.HardpointsDeployed) && !Game.Status.HasFlag(Flags.Supercruise);
+
+            VehicleMode mode;
+            if (Game.Status.HasFlag(Flags.LandingGearDeployed))
+            {
+                mode = VehicleMode.Landing;
+            }
+            else if (Game.Status.HasFlag(Flags.HudInAnalysisMode))
+            {
+                mode = VehicleMode.Analysis;
+            }
+            else
+            {
+                mode = VehicleMode.Combat;
+            }
+
+            var state = (hardpointsDeployed, mode);
+
+            if (state != _lastState)
             {
                 StartAnimation();
-                _animKbdC1 = GetBackgroundColor(_lastState, Colors.KeyboardDimBrightness);
-                _animDevC1 = GetBackgroundColor(_lastState, Colors.DeviceDimBrightness);
-                _lastState = Game.ProcessState;
+                _animC1 = GetAmbientColors(_lastState.HardpointsDeployed, _lastState.Mode);
+                _lastState = state;
             }
 
             if (Animated && AnimationElapsed >= _fadeDuration)
@@ -35,40 +58,60 @@ namespace EliteChroma.Core.Layers
                 StopAnimation();
             }
 
-            _animKbdC2 = GetBackgroundColor(Game.ProcessState, Colors.KeyboardDimBrightness);
-            _animDevC2 = GetBackgroundColor(Game.ProcessState, Colors.DeviceDimBrightness);
+            _animC2 = GetAmbientColors(hardpointsDeployed, mode);
 
-            var cLogo = Game.InMainMenu ? GameColors.EliteOrange : Game.Colors.Hud;
+            var c = _animC2;
 
-            Color cKbd = Animated
-                ? PulseColor(_animKbdC1, _animKbdC2, _fadeDuration, PulseColorType.Sawtooth)
-                : _animKbdC2;
+            if (Animated)
+            {
+                c.Keyboard = PulseColor(_animC1.Keyboard, _animC2.Keyboard, _fadeDuration, PulseColorType.Sawtooth);
+                c.Device = PulseColor(_animC1.Device, _animC2.Device, _fadeDuration, PulseColorType.Sawtooth);
+                c.Logo = PulseColor(_animC1.Logo, _animC2.Logo, _fadeDuration, PulseColorType.Sawtooth);
+            }
 
-            Color cDev = Animated
-                ? PulseColor(_animDevC1, _animDevC2, _fadeDuration, PulseColorType.Sawtooth)
-                : _animDevC2;
+            canvas.Keyboard.Set(c.Keyboard);
+            canvas.Keypad.Set(c.Keyboard);
+            canvas.Mouse.Set(c.Device);
+            canvas.Mousepad.Set(c.Device);
+            canvas.Headset.Set(c.Device);
+            canvas.ChromaLink.Set(c.Device);
 
-            canvas.Keyboard.Set(cKbd);
-            canvas.Mouse.Set(cDev);
-            canvas.Mousepad.Set(cDev);
-            canvas.Keypad.Set(cKbd);
-            canvas.Headset.Set(cDev);
-            canvas.ChromaLink.Set(cDev);
             var k = canvas.Keyboard;
-            k[Key.Logo] = cLogo;
+            k[Key.Logo] = c.Logo;
         }
 
-        private Color GetBackgroundColor(GameProcessState state, double brightness)
+        private AmbientColors GetAmbientColors(bool hardpointsDeployed, VehicleMode mode)
         {
-            switch (state)
+            Color c;
+            switch (mode)
             {
-                case GameProcessState.InForeground:
-                    return Game.Colors.Hud.Transform(brightness);
-                case GameProcessState.InBackground:
-                    return Game.Colors.Hud;
+                case VehicleMode.Combat:
+                    c = Game.Colors.Hud;
+                    break;
+                case VehicleMode.Analysis:
+                    c = Game.Colors.AnalysisMode;
+                    break;
+                case VehicleMode.Landing:
                 default:
-                    return Color.Black;
+                    c = Colors.LandingMode;
+                    break;
             }
+
+            return new AmbientColors
+            {
+                Keyboard = c.Transform(Colors.KeyboardDimBrightness),
+                Device = hardpointsDeployed ? c : c.Transform(Colors.DeviceDimBrightness),
+                Logo = Game.InMainMenu ? GameColors.EliteOrange : c,
+            };
+        }
+
+        private struct AmbientColors
+        {
+            public Color Keyboard { get; set; }
+
+            public Color Device { get; set; }
+
+            public Color Logo { get; set; }
         }
     }
 }