+++ CHANGELOG.md
@@ -9,10 +9,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Detect Epic Games installation folder
+- Add transition when changing vehicle modes
+- Add "landing gear deployed" ambient mode
 
 ### Changed
 
 - Increase animation rate to 30 fps.
+- Adjust rendering order of some effects
 
 ## [1.10.0](https://github.com/poveden/EliteChroma/compare/v1.9.2...v1.10.0) — 2021-01-10